Refactoring Validation, Email and Validator
